# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SorobanPay Backend API
  description: |
    REST API for the SorobanPay recurring-payment protocol.

    Merchants authenticate via a SEP-10 style challenge-response flow:
    1. `GET /api/v1/auth/challenge?account=G…` — obtain unsigned challenge XDR
    2. Sign the returned transaction with Freighter (or your private key)
    3. `POST /api/v1/auth/token { transaction: "<signed-XDR>" }` — exchange for JWT
    4. Include `Authorization: Bearer <token>` on all protected endpoints

    Rate limit: **100 requests / minute per IP** (HTTP 429 on breach).

    ## Structured error codes (#1074)

    All 4xx/5xx responses return an `ApiError` envelope:
    ```json
    {
      "error": "Human-readable message safe for clients",
      "code": "VALIDATION_ERROR",
      "correlationId": "req_01HXK8MZAB",
      "details": null
    }
    ```
    Stable machine-readable `code` values: `VALIDATION_ERROR`, `UNAUTHORIZED`,
    `FORBIDDEN`, `NOT_FOUND`, `RATE_LIMITED`, `INTERNAL_ERROR`,
    `DEPENDENCY_UNAVAILABLE`, `IDEMPOTENCY_CONFLICT`.
    Use `correlationId` when filing support requests or tracing logs.
  version: 1.0.0
  contact:
    name: SorobanPay
    url: https://github.com/Chrisland58/SorobanPay
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.sorobanpay.example.com
    description: Production (placeholder)

# ─── Security ──────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /api/v1/auth/token`. Include as `Authorization: Bearer <token>`.
        Tokens expire after 24 hours.
  schemas:
...
    Address:
      type: string
      pattern: "^G[A-Z2-7]{55}$"
      example: "GABC1234MERCHANT5678EXAMPLE000000000000000000000000000XYZ"
...
    ApiError:
      type: object
      required: [error, code, correlationId]
      properties:
        error:
          type: string
          description: Human-readable error message safe for clients
          example: "Merchant address is missing"
        code:
          type: string
          description: Stable machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - IDEMPOTENCY_CONFLICT
          example: "VALIDATION_ERROR"
        correlationId:
          type: string
          description: Request correlation ID for log tracing — include when filing support requests
          example: "req_01HXK8MZAB"
        details:
          type: object
          nullable: true
          description: Optional structured details (field errors, retry hints, existing request ID)
          additionalProperties: true
...
paths:
...
  /api/v1/auth/challenge:
    get:
      operationId: getAuthChallenge
      summary: Request a SEP-10 authentication challenge
      tags: [Auth]
      parameters:
        - name: account
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Address"
          description: Merchant Stellar public key (G…)
      responses:
        "200":
          description: Unsigned challenge transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    type: string
                    description: Base64-encoded unsigned Stellar transaction XDR
                  network_passphrase:
                    type: string
                  expires_in:
                    type: integer
                    description: Seconds until challenge expires
        "400":
          description: Invalid account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/auth/token:
    post:
      operationId: postAuthToken
      summary: Exchange signed challenge for a JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transaction]
              properties:
                transaction:
                  type: string
                  description: Base64-encoded signed Stellar transaction XDR
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_in:
                    type: integer
        "400":
          description: Missing transaction body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Subscriptions ────────────────────────────────────────────────────────────
...